+++ src/Raven.CodeAnalysis/Symbols/PE/PEMethodSymbol.cs
@@ -17,6 +17,7 @@ internal partial class PEMethodSymbol : PESymbol, IMethodSymbol
     private ImmutableArray<ITypeSymbol>? _typeArguments;
     private Accessibility? _accessibility;
     private ImmutableArray<IMethodSymbol>? _explicitInterfaceImplementations;
+    private ImmutableArray<AttributeData>? _attributes;
 
     public PEMethodSymbol(ReflectionTypeLoader reflectionTypeLoader, MethodBase methodInfo, INamedTypeSymbol? containingType, Location[] locations, ISymbol? associatedSymbol = null, bool addAsMember = true)
         : base(containingType, containingType, containingType.ContainingNamespace, locations, addAsMember: addAsMember)
@@ -223,6 +224,32 @@ internal partial class PEMethodSymbol : PESymbol, IMethodSymbol
 
     public ImmutableArray<AttributeData> GetReturnTypeAttributes() => ImmutableArray<AttributeData>.Empty;
 
+    public override ImmutableArray<AttributeData> GetAttributes()
+    {
+        if (_attributes.HasValue)
+            return _attributes.Value;
+
+        try
+        {
+            var builder = ImmutableArray.CreateBuilder<AttributeData>();
+
+            foreach (var attribute in _methodInfo.GetCustomAttributesData())
+            {
+                var data = TryCreateAttributeData(attribute);
+                if (data is not null)
+                    builder.Add(data);
+            }
+
+            _attributes = builder.ToImmutable();
+        }
+        catch
+        {
+            _attributes = ImmutableArray<AttributeData>.Empty;
+        }
+
+        return _attributes.Value;
+    }
+
     public override Accessibility DeclaredAccessibility => _accessibility ??= MapAccessibility(_methodInfo);
 
     public override bool IsStatic => _methodInfo.IsStatic;
@@ -325,6 +352,55 @@ internal partial class PEMethodSymbol : PESymbol, IMethodSymbol
         }
 
         return false;
+    }
+
+    private AttributeData? TryCreateAttributeData(CustomAttributeData attribute)
+    {
+        if (_reflectionTypeLoader.ResolveType(attribute.AttributeType) is not INamedTypeSymbol attributeClass)
+            return null;
+
+        var attributeConstructor = attributeClass.GetMembers(".ctor")
+            .OfType<IMethodSymbol>()
+            .FirstOrDefault(ctor => ctor.Parameters.Length == attribute.ConstructorArguments.Count)
+            ?? attributeClass.GetMembers(".ctor").OfType<IMethodSymbol>().FirstOrDefault();
+
+        if (attributeConstructor is null)
+            return null;
+
+        var constructorArguments = ImmutableArray.CreateRange(attribute.ConstructorArguments.Select(CreateTypedConstant));
+        var namedArguments = ImmutableArray.CreateRange(
+            attribute.NamedArguments.Select(named =>
+                new KeyValuePair<string, TypedConstant>(named.MemberName, CreateTypedConstant(named.TypedValue))));
+
+        return new AttributeData(
+            attributeClass,
+            attributeConstructor,
+            constructorArguments,
+            namedArguments,
+            applicationSyntaxReference: null);
+    }
+
+    private TypedConstant CreateTypedConstant(CustomAttributeTypedArgument argument)
+    {
+        var type = _reflectionTypeLoader.ResolveType(argument.ArgumentType);
+
+        if (argument.Value is null)
+            return TypedConstant.CreateNull(type);
+
+        if (argument.ArgumentType.IsArray &&
+            argument.Value is IReadOnlyCollection<CustomAttributeTypedArgument> elements)
+        {
+            var values = ImmutableArray.CreateRange(elements.Select(CreateTypedConstant));
+            return TypedConstant.CreateArray(type, values);
+        }
+
+        if (argument.Value is Type typeValue &&
+            _reflectionTypeLoader.ResolveType(typeValue) is ITypeSymbol resolvedType)
+        {
+            return TypedConstant.CreateType(type, resolvedType);
+        }
+
+        return TypedConstant.CreatePrimitive(type, argument.Value);
     }
 
     public bool IsExtern => _methodInfo.IsAbstract || (_methodInfo.Attributes & MethodAttributes.PinvokeImpl) != 0;